refactor notifications policy flow
